+++ packages/nimble-components/src/label-provider/core/label-tokens.ts
@@ -15,3 +15,18 @@ export const numericIncrementLabel = DesignToken.create<string>({
     name: 'numeric-increment-label',
     cssCustomPropertyName: null
 }).withDefault(coreLabelDefaults.numericIncrementLabel);
+
+export const errorIconLabel = DesignToken.create<string>({
+    name: 'error-icon-label',
+    cssCustomPropertyName: null
+}).withDefault(coreLabelDefaults.errorIconLabel);
+
+export const warningIconLabel = DesignToken.create<string>({
+    name: 'warning-icon-label',
+    cssCustomPropertyName: null
+}).withDefault(coreLabelDefaults.warningIconLabel);
+
+export const informationIconLabel = DesignToken.create<string>({
+    name: 'information-icon-label',
+    cssCustomPropertyName: null
+}).withDefault(coreLabelDefaults.informationIconLabel);