+++ packages/nimble-components/src/table-column/enum-text/tests/table-column-enum-text.stories.ts
@@ -1,9 +1,6 @@
 import { html, ref } from '@microsoft/fast-element';
 import type { Meta, StoryObj } from '@storybook/html';
-import {
-    createUserSelectedThemeStory,
-    incubatingWarning
-} from '../../../utilities/tests/storybook';
+import { createUserSelectedThemeStory } from '../../../utilities/tests/storybook';
 import { tableTag } from '../../../table';
 import { tableColumnEnumTextTag } from '..';
 import {
@@ -44,7 +41,7 @@ const enumTextColumnDescription = `The \`nimble-table-column-enum-text\` column 
 ${columnOperationBehavior}`;
 
 const metadata: Meta<EnumTextColumnTableArgs> = {
-    title: 'Incubating/Table Column - Enum Text',
+    title: 'Components/Table Column: Enum Text',
     parameters: {
         docs: {
             description: {
@@ -66,7 +63,6 @@ interface EnumTextColumnTableArgs extends SharedTableArgs {
 export const enumTextColumn: StoryObj<EnumTextColumnTableArgs> = {
     // prettier-ignore
     render: createUserSelectedThemeStory(html<EnumTextColumnTableArgs>`
-        ${incubatingWarning({ componentName: 'table', statusLink: 'https://github.com/orgs/ni/projects/7/views/21' })}
         <${tableTag}
             ${ref('tableRef')}
             data-unused="${x => x.updateData(x)}"